rearrange zep modules
